+++ internal/api/budget_test.go
@@ -6,9 +6,8 @@ import (
 	"encoding/json"
 
 	"github.com/stretchr/testify/assert"
-	// "github.com/stretchr/testify/require"
 
-	"purch/database"
+	"purch/internal/database"
 )
 
 // login credentials
